requirements.txt
aiodns==3.0.0
aiofiles=23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.2.0
aiosocksy==0.1.2
anyio==3.6.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==23.1.2
autocommand==2.2.1
Automat==22.10.0
Babel==2.9.1
binance-futures-connector==3.3.1
Brotli==1.0.9
ccxt==3.0.74
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.1
cheroot==8.6.0
CherryPy==18.8.0
click==8.1.3
common==0.1.2
constantly==15.1.0
cryptography==38.0.1
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
Django==4.1.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
dnspython==2.2.1
email-validator==1.2.1
exceptiongroup==1.0.4
fake-useragent==1.1.1
Flask==2.2.3
frozenlist==1.3.1
gunicorn==20.0.4
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
hypercorn==0.14.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
inflect==6.0.0
iniconfig==1.1.1
ipaddress==1.0.23
itsdangerous==2.1.2
jaraco.classes==3.2.2
jaraco.collections==3.5.2
jaraco.context==4.1.2
jaraco.functools==3.5.1
jaraco.text==3.9.1
Jinja2==3.1.2
loguru==0.6.0
lxml==4.9.1
magic-filter=1.0.9
MarkupSafe==2.1.2
more-itertools==8.14.0
multidict==6.0.2
multitasking==0.0.11
numpy==1.24.3
packaging==21.3
pandas==2.0.1
phonenumbers==8.12.55
plotly==5.14.1
pluggy==1.0.0
portend==3.1.0
priority==2.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.3.0
pycares==4.2.2
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.9
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyQiwiP2P==2.0.6
PySocks==1.7.1
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-socks==2.3.0
pytz==2022.2.1
requests==2.28.1
response-report==1.1.1
rfc3986==1.5.0
rsa==4.9
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sqlparse==0.4.2
starlette==0.20.4
Telethon==1.27.0
tempora==5.0.2
tenacity==8.2.2
toml==0.10.2
tomli==2.0.1
tradingview-ta==3.3.0
Twisted==22.10.0
txaio==23.1.1
typing==3.7.4.3
typing_extensions==4.3.0
tzdata==2023.3
ujson==5.4.0
urllib3==1.26.12
validators==0.20.0
websockets==10.3
Werkzeug==2.2.3
wsproto==1.2.0
yarl==1.8.1
zc.lockfile==2.0
zope.interface==6.0
